Improved spritesheet loading.
git-svn-id: https://dev.wibe.media/svn/code/java/trunk/litiengine@2260 16493bb7-3577-b145-b9c3-72a97d279fb6

+++ src/de/gurkenlabs/litiengine/graphics/particles/RectangleFillParticle.java
@@ -21,6 +21,6 @@ public class RectangleFillParticle extends Particle {
   public void render(final Graphics2D g, final Point2D emitterOrigin) {
     final Point2D renderLocation = this.getLocation(emitterOrigin);
     g.setColor(this.getColor());
-    g.fill(new Rectangle2D.Float((float) renderLocation.getX(), (float) renderLocation.getY(), this.getWidth(), this.getHeight()));
+    g.fill(new Rectangle2D.Float((float) renderLocation.getX(), (float) renderLocation.getY(), (int)this.getWidth(), (int)this.getHeight()));
   }
 }
